reworking data models for development

+++ db/migrate/20110119202133_create_proposals.rb
@@ -2,7 +2,7 @@ class CreateProposals < ActiveRecord::Migration
   def self.up
     create_table :proposals do |t|
       t.string :name
-      t.float :priority
+      t.decimal :priority
 
       t.timestamps
     end
